% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{device_clear}
\alias{device_clear}
\title{Clear an object associated with the window.}
\usage{
device_clear(clear_types = c("shapes", "bboxdeco", "lights", "background"))
}
\arguments{
\item{clear_types}{A character vector that specifies the types of
objects to clear. Accepted values for the vector include: "shapes", "bboxdeco", "lights", "background".
The default is all 4 types are specified.}
}
\description{
Function clears specific types of objects from the device window including "shapes", "bboxdeco", "lights", "background".
}
