% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{device_close_current}
\alias{device_close_current}
\title{Close the device window.}
\usage{
device_close_current()
}
\value{
Returns the current list of device id's
}
\description{
Function closes the current device window.
See 'device_current()' function for setting the current window.
}
