% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{device_current}
\alias{device_current}
\title{Set or return the integer id of the current window.}
\usage{
device_current(id)
}
\arguments{
\item{id}{If defined then sets the current device window to 'id'.}
}
\value{
If id is undefined, returns the current device id among one or more device windows.
}
\description{
Function is used to set a specific window as 'current' or return the id of the 'current' window.
In setting the id, the value is checked for existence among the list of window id's.
}
