% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{device_list}
\alias{device_list}
\title{Get a list of all device window ids.}
\usage{
device_list()
}
\value{
Returns a list of all the integer ids of the created windows.
}
\description{
Function returns a list of all created windows with their respective device id's.
}
