% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{device_new}
\alias{device_new}
\title{Create a new 3d device window.}
\usage{
device_new(
  bgrd_color = "white",
  width = 640,
  height = 640,
  vp_theta = 0,
  vp_phi = 15,
  vp_fov = 60,
  vp_zoom = 1
)
}
\arguments{
\item{bgrd_color}{The String background color of the window. Default is "white".}

\item{width}{The integer width of the window in pixels. The default is 640.}

\item{height}{The integer height of the window in pixels. The default is 640.}

\item{vp_theta}{The rotation of the camera. The default is 0.}

\item{vp_phi}{The rotation of the camera. The default is 15.}

\item{vp_fov}{The field of view angle. The default is 60.}

\item{vp_zoom}{The degree of zoom for the camera. The default is 1.}
}
\value{
The newly created integer id for the device window.
}
\description{
Function creates a new window with options for background color, size, camera angles, field of view angle,
and zoom.
}
