% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot}
\alias{dot}
\title{Function computes the dot product (inner product) of two 3x1 matrices}
\usage{
dot(v1, v2)
}
\arguments{
\item{v1}{A numeric vector or 3x1 numeric matrix}

\item{v2}{A numeric vector or 3x1 numeric matrix}
}
\value{
A numeric value
}
\description{
Function accepts two numeric vectors or two 3x1 matrices
}
\author{
Rick Dean
}
