% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_norm.R
\name{E_norm}
\alias{E_norm}
\title{Function returns the length of a vector or column lengths from a matrix}
\usage{
E_norm(X)
}
\arguments{
\item{X}{a numeric vector or matrix}
}
\value{
a scalar norm (if \code{X} is a vector) or a vector (if \code{X} is a matrix)
of norms for each column.
}
\description{
Function calculates the Euclidean norm of a vector
or norm for each column if a matrix is submitted.
}
\author{
Rick Dean
}
