% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.square.R
\name{is.square}
\alias{is.square}
\title{Function returns TRUE if matrix is square}
\usage{
is.square(x)
}
\arguments{
\item{x}{A matrix or data frame that can be converted to a matrix.}
}
\value{
Returns TRUE if argument \code{x} is square
}
\description{
Function returns TRUE if matrix is square
}
