% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_rank.R
\name{ls_rank}
\alias{ls_rank}
\title{Function returns the rank of a linear system Ax = b and determines if
the system has a unique solution, an infinite number of solutions, or
no solution.}
\usage{
ls_rank(A, b)
}
\arguments{
\item{A}{A numeric matrix or data frame that represents a system of equations.}

\item{b}{A numeric vector or matrix that represents the constant values \code{b}
in \code{Ax = b}.}
}
\value{
Function returns a named list with the ranks of both \code{A} and the
augmented matrix \code{A|b}, and whether the solution is "unique", "infinite", or
"none".
}
\description{
Function returns the rank of a linear system Ax = b and determines if
the system has a unique solution, an infinite number of solutions, or
no solution.
}
