% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{object_list}
\alias{object_list}
\title{Get a list of objects in the device window.}
\usage{
object_list()
}
\value{
A dataframe containing all the object ids' for the current device window.
}
\description{
Function returns a list of all the object id's contained in the current window.
}
