% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{object_remove}
\alias{object_remove}
\title{Remove an object from the window.}
\usage{
object_remove(object_id)
}
\arguments{
\item{object_id}{An object's specific integer id that selects and removes it from the scene.}
}
\description{
Function removes a specific object from the window.  The submitted object id is checked for
existence among the windows listed id's.
}
