% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_equations_3d.R
\name{plot_equations_3d}
\alias{plot_equations_3d}
\title{Function plots linear equations in 3d space.}
\usage{
plot_equations_3d(
  win_width = 1000,
  win_height = 800,
  A,
  b,
  equation_color = "red",
  alpha = 0.5,
  lit = TRUE,
  bkg_color = "white",
  draw_axis = TRUE,
  main_lab = NULL,
  x_lab = "X",
  y_lab = "Y",
  z_lab = "Z",
  axis_color = "black",
  xlim = c(0, 10),
  ylim = c(0, 10),
  zlim = c(0, 10),
  html_file = NULL
)
}
\arguments{
\item{win_width, win_height}{An integer defining the dimensions of the plot window.}

\item{A}{Either the matrix of coefficients of a system of linear equations, or a data frame with both \code{A}
and \code{b} as combined columns.}

\item{b}{If supplied, the vector of constants on the right hand side of the equations, of length matching
the number of rows of \code{A}.}

\item{equation_color}{A scalar or vector that defines each equation's plane fill color.}

\item{alpha}{A numeric that defines the plane color's alpha component.}

\item{lit}{logical, specifying if lighting calculation should take place on geometry; see \code{rgl::rgl.material}}

\item{bkg_color}{A character string defining the window background color.}

\item{draw_axis}{A logical which if TRUE draws the axes.}

\item{main_lab}{A character string defining the plot's main title.}

\item{x_lab, y_lab, z_lab}{A character string defining the axis labels.}

\item{axis_color}{A character string defining the axis color.}

\item{xlim, ylim, zlim}{An integer vector defining the min/max of an axis.}

\item{html_file}{A character string that defines the file path to creating an html of the plot.}
}
\value{
The device window id (an integer).
}
\description{
Function plots one or more system of linear equations in the form \eqn{A x = b} with three unknowns,
x1, x2, and x3, by plotting a plane for each equation.
}
\author{
Rick Dean
}
