% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines_3d.R
\name{plot_lines_3d}
\alias{plot_lines_3d}
\title{Function plots line segments in a 3d device window}
\usage{
plot_lines_3d(
  segments,
  win_width = 1000,
  win_height = 800,
  segment_colors = NULL,
  segment_widths = NULL,
  bkg_color = "white",
  draw_axis = TRUE,
  main_title = NULL,
  x_lab = "X",
  y_lab = "Y",
  z_lab = "Z",
  axis_color = "black",
  xlim = c(0, 10),
  ylim = c(0, 10),
  zlim = c(0, 10),
  x_ticks = 11,
  y_ticks = 11,
  z_ticks = 11,
  html_file = NULL
)
}
\arguments{
\item{segments}{A numeric matrix where each row is a line segment defined by a pair of points in 3d space.
The first three columns defines the x/y/z coordinates for the first point; the next three columns define
the second point of the segment.}

\item{win_width, win_height}{An integer defining the dimensions of the plot window.}

\item{segment_colors}{A vector of strings defining each segment's color. The default is "black".
The length of \code{segment_colors} should be the same as \code{segments}.}

\item{segment_widths}{A numeric vector defining each segment's width. The default is 1.0.
The length of \code{segment_widths} should be the same as \code{segments}.}

\item{bkg_color}{A character string defining the window background color.}

\item{draw_axis}{A logical which if TRUE draws the axes.}

\item{main_title}{A character string defining the plot's main title.}

\item{x_lab, y_lab, z_lab}{A character string defining the axis labels.}

\item{axis_color}{A character string defining the axis color.}

\item{xlim, ylim, zlim}{An integer vector defining the min/max of an axis.}

\item{x_ticks, y_ticks, z_ticks}{An integer defining the number of ticks for an axis.}

\item{html_file}{A character string that defines the file path for creating
a WebGL version of the plot. File names should have an html extension.}
}
\value{
The device window id (an integer).
}
\description{
Function plots a collection of line segments where \code{segments} is a numeric matrix of six
columns of start and end points in 3d space. Options are provided for sizing the device window,
width/color of the segments, and defining/labeling axis,
}
