% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points_3d.R
\name{plot_points_3d}
\alias{plot_points_3d}
\title{Function plot points in a 3d device window}
\usage{
plot_points_3d(
  df,
  x_col,
  y_col,
  z_col,
  group_col = NULL,
  win_width = 1000,
  win_height = 800,
  pt_colors = "black",
  pt_size = 6,
  pt_alpha = 1,
  bkg_color = "white",
  draw_axis = TRUE,
  main_title = NULL,
  x_lab = "X",
  y_lab = "Y",
  z_lab = "Z",
  axis_color = "black",
  xlim = c(0, 10),
  ylim = c(0, 10),
  zlim = c(0, 10),
  x_ticks = 11,
  y_ticks = 11,
  z_ticks = 11,
  html_file = NULL
)
}
\arguments{
\item{df}{A data frame  with columns for the x, y, and z locations and optionally a column for grouping the points.}

\item{x_col, y_col, z_col}{Strings that set the columns names from \code{df} for the x, y, z numeric coordinates.}

\item{group_col}{An optional string that sets the column name from \code{df} for grouping the points.  The
column must be a factor with discrete levels.}

\item{win_width, win_height}{An integer defining the dimensions of the plot window.}

\item{pt_colors}{Can be single string that sets the color for all the points. The default is "black". If \code{group_col}
has been assigned then this parameter is a vector that sets the colors for each of factor column's levels.}

\item{pt_size}{A numeric defining the size of all the points in pixels. The default is 6.0.}

\item{pt_alpha}{A numeric between 0.0 and 1.0 that sets the alpha value for all the points. The default is 1.0.}

\item{bkg_color}{A character string defining the window background color. The default is "white".}

\item{draw_axis}{A logical which if TRUE draws the axes.}

\item{main_title}{A character string defining the plot's main title.}

\item{x_lab, y_lab, z_lab}{A character string defining the axis labels.}

\item{axis_color}{A character string defining the axis color. The default is "black".}

\item{xlim, ylim, zlim}{An integer vector defining the min/max of an axis.}

\item{x_ticks, y_ticks, z_ticks}{An integer defining the number of ticks for an axis.}

\item{html_file}{A character string that defines the file path for creating
a WebGL version of the plot. File names should have an html extension.}
}
\value{
The device window id (an integer).
}
\description{
Function plots the 3d points in the current device window.  A data frame provides the x, y, and z
location of the points.  Options for point color and size along with labeling and axis options are provided.
}
