% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vector_2d.R
\name{plot_vector_2d}
\alias{plot_vector_2d}
\title{Function plots 2d arrowed vectors}
\usage{
plot_vector_2d(
  from_to = NULL,
  to = NULL,
  vector_labels = NULL,
  vector_colors = NULL,
  vector_widths = NULL,
  vector_linetypes = NULL,
  draw_axis = TRUE,
  axis_color = "black",
  grid_color = "lightgreen",
  lines_00_color = "darkgreen",
  x_breaks = c(-5, 5, 1),
  y_breaks = c(-5, 5, 1),
  x_title = "X",
  y_title = "Y",
  width_height = c(6, 6),
  char_mag = 1.5
)
}
\arguments{
\item{from_to}{A n x 4 numeric matrix that defines n vectors where the first two columns
define the x/y "from" point and the last two columns define the x/y "to" point of each vector.}

\item{to}{An n x 2 numeric matrix that defines n vectors where the x/y "from" point is assumed to be
at the origin \code{(0,0)} and the two columns of \code{to} define the x/y "to" point for each vector.}

\item{vector_labels}{A vector of strings with the same length as \code{from_to} or \code{to} that defines a label for each vector.}

\item{vector_colors}{A vector of strings with the same length as \code{to} or \code{from_to} that defines a color for each vector.}

\item{vector_widths}{A numeric vector with the same length as \code{to} or \code{from_to} that defines a width for each vector.}

\item{vector_linetypes}{A vector of strings with the same length as \code{to} or \code{from_to} that defines a line type for each vector.
Acceptable values are "blank", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash".}

\item{draw_axis}{A logical which if TRUE, draws the numeric x/y axis'.}

\item{axis_color}{A string that defines the axis color.}

\item{grid_color}{A string that defines the grid color.}

\item{lines_00_color}{A string that defines the color for vertical/horizontal lines from (0,0).}

\item{x_breaks}{A numeric three element vector of upper, lower, and interval for exact x axis tic locations.}

\item{y_breaks}{A numeric three element vector of upper, lower, and interval for exact y axis tic locations.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{width_height}{A two element numeric vector that defines the plot's x/y dimensions in inches.}

\item{char_mag}{A numeric that defines the magnification of the vector's label size.}
}
\description{
Function plots one or more arrowed vectors across a 2d plotting space. Options are provided to add an x-y axis
and a grid. Either the \code{from_to} or \code{to} arguments may be specified to populate the plot.
}
