% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_vector.R
\name{proj_vector}
\alias{proj_vector}
\title{Function computes the projection of a vector onto another vector}
\usage{
proj_vector(x, v)
}
\arguments{
\item{x}{A numeric vector that is to be projected onto \code{v}}

\item{v}{A numeric vector}
}
\value{
The coordinates of a projected vector onto another vector..
}
\description{
Function returns the vector coordinates of a vector projected
onto another vector.
}
\author{
Rick Dean
}
