% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_gs.R
\name{qr_gs}
\alias{qr_gs}
\title{Function produces an orthogonal basis \code{Q} for a vector space represented
by a numeric matrix \code{X} via Gram-Schmidt}
\usage{
qr_gs(x)
}
\arguments{
\item{x}{A numeric matrix or data frame}
}
\value{
A numeric matrix of orthogonal basis column vectors
}
\description{
Function uses the Gram-Schmidt process to convert the
columns of a basis matrix \code{X} into an orthogonal basis \code{Q}. The
function also returns an upper triangular matrix \code{R} where \code{X = QR},
known as QR factorization.
}
\author{
Rick Dean
}
