% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_householder.R
\name{qr_householder}
\alias{qr_householder}
\title{Function performs a QR decomposition of numeric matrix by converting it
to a Householder matrix or reflector matrix.}
\usage{
qr_householder(x, complete = FALSE)
}
\arguments{
\item{x}{A numeric matrix or data frame}

\item{complete}{A logical which it TRUE returns a "thin" orthogonal matrix Q
and triangular matrix R.}
}
\value{
A named list with the orthogonal matrix Q and triangular matrix R.
}
\description{
For a numeric matrix X of n x m  (m <= n) then the Householder
process rotates/reflects into an orthogonal matrix Q. The function also
returns an upper triangular matrix \code{R} where \code{X = QR},
known as QR factorization.
}
\author{
Rick Dean
}
