% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rref.R
\name{rref}
\alias{rref}
\title{Function implements Gauss-Jordan elimination algorithm in converting a numeric matrix to
reduced row-echelon form (rref).}
\usage{
rref(x)
}
\arguments{
\item{x}{a numeric matrix or data frame.}
}
\value{
A matrix in reduced reduced row echelon form.
}
\description{
Function implements Gauss-Jordan elimination algorithm in converting a numeric matrix to
reduced row-echelon form (rref).
}
\author{
Rick Dean
}
