% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Function returns the trace of a square numerical matrix}
\usage{
tr(x)
}
\arguments{
\item{x}{A square numeric matrix or data frame that can be
converted to a square numeric matrix.}
}
\value{
A scalar that represents the trace of a matrix
}
\description{
Function returns the trace of a square numerical matrix
}
