% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_latex.R
\name{matrix_to_latex}
\alias{matrix_to_latex}
\title{Function converts numeric/character matrix or data frame to Latex format}
\usage{
matrix_to_latex(
  x = NULL,
  bracket_type = "bracket",
  small_inline = FALSE,
  fractions = FALSE,
  digits = 2,
  matrix_name = NULL,
  centered = FALSE
)
}
\arguments{
\item{x}{A numeric/character vector, matrix, data frame.}

\item{bracket_type}{A string that set the bracket type. Acceptable values are
"none", "bracket", "paren", "curly_bracket", "verts",
"double_verts".}

\item{small_inline}{A logical which if \code{TRUE} creates an inline small matrix.}

\item{fractions}{A logical which if \code{TRUE} will express non-integers
as rational numbers.}

\item{digits}{The number of rounded digits for the numeric values.}

\item{matrix_name}{A string that sets the name of the matrix.}

\item{centered}{A logical which if \code{TRUE} creates Rmarkdown that centers the matrix.}
}
\value{
A string of inline Rmarkdown
}
\description{
Function returns a Latex formatted matrix for Rmarkdown documents.

Create the Latex string by calling the function:
\code{A_matrix_str = matrix_to_latex(A)}. The matrix string could then
be used directly in a Rmarkdown document, surrounding latex "$" signs included.
}
\author{
Rick Dean
}
